+++ packages/core/src/targets.ts
@@ -1,101 +1,99 @@
+import SetMap from './data_structures/set_map';
 import type ImpulseElement from './element';
 import { getAttributeValues } from './helpers/dom';
 import AttributeObserver from './observers/attribute_observer';
+import Scope from './scope';
 import Store from './store';
-
-const ATTRIBUTE_NAME = 'data-target';
 
 export default class Targets {
   private store: Store;
+  private scope: Scope;
+  private targetsByName: SetMap<string, Element>;
   private attributeObserver: AttributeObserver;
 
   constructor(private readonly instance: ImpulseElement) {
     this.instance = instance;
     this.store = new Store(Object.getPrototypeOf(this.instance), 'targets');
-    this.attributeObserver = new AttributeObserver(this.instance, ATTRIBUTE_NAME, this);
+    this.scope = new Scope(this.instance);
+    this.targetsByName = new SetMap();
+    this.attributeObserver = new AttributeObserver(this.instance, 'data-target', this);
   }
 
   start() {
     this.attributeObserver.start();
-    this.targetKeys.forEach((key) => this.initializeKey(key));
+    this.keys.forEach((key) => this.initializeKey(key));
   }
 
   stop() {
-    this.targetKeys.forEach((key) => this.terminateKey(key));
+    this.keys.forEach((key) => this.terminateKey(key));
     this.attributeObserver.stop();
   }
 
-  elementConnected(element: Element) {
-    const values = getAttributeValues(element, ATTRIBUTE_NAME);
-    values.forEach((value) => {
-      const [elementName, key] = value.split('.');
-      if (!elementName || !key || elementName !== this.identifier) return;
-      // Element needs to be registered before we can process its addition.
-      if (!this.targetKeys.includes(key)) return;
-      // In the callback function, we should be able to call `this.elements` and get back all the elements. Therefore,
-      // we need to set the property before calling the function.
-      this.defineProperty(key, this.findTargets(key));
-      this.processAddedElement(key, element);
-    });
+  elementConnected(target: Element) {
+    const tokens = getAttributeValues(target, 'data-target');
+    for (const token of tokens) {
+      const [identifier, key] = token.split('.');
+      if (
+        key &&
+        !this.targetsByName.has(key, target) &&
+        identifier === this.identifier &&
+        this.keys.includes(key) &&
+        this.findTargets(key).includes(target)
+      ) {
+        this.targetsByName.add(key, target);
+        this.defineProperty(key, this.findTargets(key));
+        this.invokeConnectedCallback(key, target);
+      }
+    }
   }
 
-  elementDisconnected(element: Element) {
-    const values = getAttributeValues(element, ATTRIBUTE_NAME);
-    values.forEach((value) => {
-      const [elementName, key] = value.split('.');
-      if (!elementName || !key || elementName !== this.identifier) return;
-      // Element needs to be registered before we can process its removal.
-      if (!this.targetKeys.includes(key)) return;
-      this.processRemovedElement(key, element);
-      // In the callback function, we should be able to call `this.elements` and get back all the elements. Therefore,
-      // we need to set the property after calling the function. If you need an updated list of elements within this
-      // function, consider running the code after `await new Promise((resolve) => requestAnimationFrame(resolve))`.
-      this.defineProperty(key, this.findTargets(key));
-    });
+  elementDisconnected(target: Element) {
+    const tokens = getAttributeValues(target, 'data-target');
+    for (const token of tokens) {
+      const [identifier, key] = token.split('.');
+      if (key && this.targetsByName.has(key, target) && identifier === this.identifier && this.keys.includes(key)) {
+        this.invokeDisconnectedCallback(key, target);
+        this.defineProperty(key, this.findTargets(key));
+      }
+    }
   }
 
   private initializeKey(key: string) {
-    // If the key has not been registered, return.
-    if (!this.targetKeys.includes(key)) return;
-
     const targets = this.findTargets(key);
+    targets.forEach((target) => this.targetsByName.add(key, target));
     this.defineProperty(key, targets);
-    targets.forEach((target) => this.processAddedElement(key, target));
+    targets.forEach((target) => this.invokeConnectedCallback(key, target));
   }
 
   private terminateKey(key: string) {
-    // If the key has not been registered, return.
-    if (!this.targetKeys.includes(key)) return;
-
     const targets = this.findTargets(key);
-    targets.forEach((target) => this.processRemovedElement(key, target));
-  }
-
-  private defineProperty(key: string, value: Element[]) {
-    const descriptor: PropertyDescriptor = { configurable: true, get: () => value };
-    Object.defineProperty(this.instance, key, descriptor);
-  }
-
-  private processAddedElement(key: string, element: Element) {
-    const fn = (this.instance as unknown as Record<string, unknown>)[`${key}Connected`];
-    if (typeof fn === 'function') {
-      fn.call(this.instance, element);
-    }
-  }
-
-  private processRemovedElement(key: string, element: Element) {
-    const fn = (this.instance as unknown as Record<string, unknown>)[`${key}Disconnected`];
-    if (typeof fn === 'function') {
-      fn.call(this.instance, element);
-    }
+    targets.forEach((target) => this.invokeDisconnectedCallback(key, target));
   }
 
   private findTargets(key: string) {
-    const selector = `[${ATTRIBUTE_NAME}~="${this.identifier}.${key}"]`;
-    return Array.from(this.instance.querySelectorAll(selector));
+    return this.scope.findTargets(`[data-target~="${this.identifier}.${key}"]`);
   }
 
-  private get targetKeys() {
+  private defineProperty(key: string, targets: Element[]) {
+    const descriptor: PropertyDescriptor = { configurable: true, get: () => targets };
+    Object.defineProperty(this.instance, key, descriptor);
+  }
+
+  private invokeConnectedCallback(key: string, target: Element) {
+    const fn = (this.instance as unknown as Record<string, unknown>)[`${key}Connected`];
+    if (typeof fn === 'function') {
+      fn.call(this.instance, target);
+    }
+  }
+
+  private invokeDisconnectedCallback(key: string, target: Element) {
+    const fn = (this.instance as unknown as Record<string, unknown>)[`${key}Disconnected`];
+    if (typeof fn === 'function') {
+      fn.call(this.instance, target);
+    }
+  }
+
+  private get keys() {
     return Array.from(this.store.value as Set<{ key: string }>).map(({ key }) => key);
   }